# Conflicts:
#	.gitignore
#	stonebreak-game/src/main/java/com/stonebreak/core/Main.java
#	stonebreak-game/src/main/java/com/stonebreak/ui/DebugOverlay.java
#	stonebreak-game/src/main/java/com/stonebreak/world/World.java
#	stonebreak-game/src/main/java/com/stonebreak/world/spawn/SpawnLocator.java
#	stonebreak-game/src/test/java/com/stonebreak/network/bridge/VoxelChunkBulkCodecTest.java
#	stonebreak-game/src/test/java/com/stonebreak/world/generation/FusedChunkGenParityTest.java
#	stonebreak-game/src/test/java/com/stonebreak/world/generation/NativeCarverIntegrationTest.java
#	stonebreak-game/src/test/java/com/stonebreak/world/generation/noise/TerrainNoiseParityTest.java

+++ stonebreak-game/src/main/java/com/stonebreak/world/save/io/ChunkCodec.java
@@ -242,7 +242,7 @@ public final class ChunkCodec {
                 int rawLen = in.readInt();
                 int compLen = in.readInt();
                 if (rawLen <= 0 || rawLen > MAX_RAW_BYTES || compLen <= 0
-                        || HEADER_BYTES + 9 + compLen > payload.length) {
+                        || HEADER_BYTES + 9 + (long) compLen > payload.length) {
                     throw new IOException("Block section stream lengths invalid: raw="
                         + rawLen + ", compressed=" + compLen);
                 }
